examples/defi_event_generator/pyproject.toml
[project]
name = "deltastream-sdk-defi-event-generator"
version = "0.1.0"
description = "Example demonstrating DeFi event generation with flash loan attack simulation using DeltaStream SDK"
authors = [
    {name = "DeltaStream Team", email = "support@deltastream.io"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "deltastream-sdk",
]

[project.scripts]
defi-event-generator = "src.defi_event_generator:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.uv.sources]
deltastream-sdk = { path = "../../", editable = true }
